organised files, restructured file paths
